+++ tests/NginxTest.php
@@ -4,8 +4,9 @@ use Illuminate\Container\Container;
 use Valet\Configuration;
 use Valet\Filesystem;
 use Valet\Nginx;
-use function Valet\resolve;
 use Valet\Site;
+
+use function Valet\resolve;
 use function Valet\swap;
 use function Valet\user;