+++ tests/SiteTest.php
@@ -5,8 +5,9 @@ use Valet\Brew;
 use Valet\CommandLine;
 use Valet\Configuration;
 use Valet\Filesystem;
-use function Valet\resolve;
 use Valet\Site;
+
+use function Valet\resolve;
 use function Valet\swap;
 use function Valet\user;